WeAreSorryHubBase - bringing `HubBase` back to fix binary compatibility
