+++ src/Microsoft.AspNet.SignalR.Core/Hub.cs
@@ -8,68 +8,18 @@ namespace Microsoft.AspNet.SignalR
     /// <summary>
     /// Provides methods that communicate with SignalR connections that connected to a <see cref="Hub"/>.
     /// </summary>
-    public abstract class Hub : IHub
+    public abstract class Hub : HubBase
     {
-        protected Hub()
+        public IHubCallerConnectionContext<dynamic> Clients
         {
-            Clients = new HubConnectionContext();
-        }
-
-        /// <summary>
-        /// Gets a dynamic object that represents all clients connected to this hub (not hub instance).
-        /// </summary>
-        public IHubCallerConnectionContext<dynamic> Clients { get; set; }
-
-        /// <summary>
-        /// Provides information about the calling client.
-        /// </summary>
-        public HubCallerContext Context { get; set; }
-
-        /// <summary>
-        /// The group manager for this hub instance.
-        /// </summary>
-        public IGroupManager Groups { get; set; }
-
-        /// <summary>
-        /// Called when a connection disconnects from this hub gracefully or due to a timeout.
-        /// </summary>
-        /// <param name="stopCalled">
-        /// true, if stop was called on the client closing the connection gracefully;
-        /// false, if the connection has been lost for longer than the
-        /// <see cref="Configuration.IConfigurationManager.DisconnectTimeout"/>.
-        /// Timeouts can be caused by clients reconnecting to another SignalR server in scaleout.
-        /// </param>
-        /// <returns>A <see cref="Task"/></returns>
-        public virtual Task OnDisconnected(bool stopCalled)
-        {
-            return TaskAsyncHelper.Empty;
-        }
-
-        /// <summary>
-        /// Called when the connection connects to this hub instance.
-        /// </summary>
-        /// <returns>A <see cref="Task"/></returns>
-        public virtual Task OnConnected()
-        {
-            return TaskAsyncHelper.Empty;
-        }
-
-        /// <summary>
-        /// Called when the connection reconnects to this hub instance.
-        /// </summary>
-        /// <returns>A <see cref="Task"/></returns>
-        public virtual Task OnReconnected()
-        {
-            return TaskAsyncHelper.Empty;
-        }
-
-        protected virtual void Dispose(bool disposing)
-        {
-        }
-
-        public void Dispose()
-        {
-            Dispose(true);
+            get
+            {
+                return HubClients;
+            }
+            set
+            {
+                HubClients = value;
+            }
         }
     }
 }
